Almost finished with API and finished with DB

+++ Pango.Domain/Enum/PlatformEnum.cs
@@ -1,6 +1,6 @@
 ﻿namespace Pango.Domain.Enum;
 
-public enum PlatformEnum : byte
+public enum PlatformEnum : int
 {
     iOS = 1,
     Android = 2,
